+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [1.0.18] - 2023-12-13
+
+### Changed
+
+- Default to not including metrics unless specified.
+- Changed the default reportInterval to 60s
+
 ## [1.0.17] - 2023-11-27
 
 ### Changed